% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{cistromec}
\alias{cistromec}
\title{Cistrome Cancer query for ETV4 and TFAP2A}
\format{An object of class \code{data.frame} with 13279 rows and 3 columns.}
\usage{
cistromec
}
\description{
Cistrome Cancer query for ETV4 and TFAP2A
}
\references{
\url{http://cistrome.org/CistromeCancer/CancerTarget}
}
\keyword{datasets}
