% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor_igraph}
\alias{cor_igraph}
\title{Make an igraph object}
\usage{
cor_igraph(ob)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}
}
\value{
An \code{igraph} object
}
\description{
An \code{igraph} object of from \link{cmicroRNA} or \link{cTF} 
objects.
}
\examples{
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# get data for 2 microRNAs in the ACC study
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'))
DBI::dbDisconnect(conn)

# convert to cmicroRNA object and plot
ob <- cmicroRNA(dat)
g <- cor_igraph(ob)

}
