% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.tidy.R
\name{cor_prep}
\alias{cor_prep}
\title{Prepare correlation data for plotting}
\usage{
cor_prep(ob, study, add_dir = TRUE, add_corr = TRUE)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{study}{A \code{character} vector of The Cancer Genome Atlas (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
default \code{NULL} all available studies will be included.}

\item{add_dir}{A \code{logical} default TRUE for whether to add a column
called Direction that has the direction of the correlation; positive or 
negative.}

\item{add_corr}{A \code{logical} default TRUE for whether to add a column
called Correlation that has the absolute value of the correlation}
}
\value{
A \code{data.frame}
}
\description{
Not meant to be called directly by the user.
}
