% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor_print}
\alias{cor_print}
\title{Print method for \link{cmicroRNA} and \link{cTF} objects}
\usage{
cor_print(ob, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{...}{Other argument to \code{\link[base]{print}}.}
}
\value{
Printed text
}
\description{
Print method for \link{cmicroRNA} and \link{cTF} objects
}
\examples{
# load required libraries
library(RSQLite)
library(cRegulome)

# locate the testset file and connect
fl <- system.file('extdata', 'cRegulome.db', package = 'cRegulome')
conn <- dbConnect(SQLite(), fl)

# enter a custom query with different arguments
dat <- get_mir(conn,
               mir = 'hsa-let-7g',
               study = 'STES',
               min_abs_cor = .3,
               max_num = 5)

# make a cmicroRNA object   
cmir <- cmicroRNA(dat)

# print object
cor_print(cmir)

}
