% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor_tidy}
\alias{cor_tidy}
\title{Tidy \link{cmicroRNA} and \link{cTF} objects}
\usage{
cor_tidy(ob)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}
}
\value{
A tidy \code{data.frame} of four columns. \code{mirna_base} or
\code{tf}is the microRNA miRBase IDs, \code{feature} is the features/genes,
\code{cor} is the corresponding expression correlations and \code{study}
is TCGA study ID.
}
\description{
Tidy \link{cmicroRNA} and \link{cTF} objects
}
\examples{
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# get data for 2 microRNAs in the ACC study
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'),
    min_abs_cor = .5)
DBI::dbDisconnect(conn)

# convert to cmicroRNA object
ob <- cmicroRNA(dat)
dat <- cor_tidy(ob)
dat[1:5,]

}
