% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{creg_mir}
\alias{creg_mir}
\title{cRegulome query for 34 microRNAs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8487 rows and 4 columns.}
\usage{
creg_mir
}
\description{
cRegulome query for 34 microRNAs
}
\keyword{datasets}
