% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{creg_tf}
\alias{creg_tf}
\title{cRegulome query for 4 transcription factors}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12072 rows and 4 columns.}
\usage{
creg_tf
}
\description{
cRegulome query for 4 transcription factors
}
\keyword{datasets}
