% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_db}
\alias{get_db}
\title{Get cRegulome.db file}
\usage{
get_db(test = FALSE, ...)
}
\arguments{
\item{test}{A \code{logical}, default \code{FALSE}. When \code{TRUE}
downlaods a database file with the same structure with a subset of
the data for speed.}

\item{...}{Optional arguments passed to \code{\link[utils]{download.file}}}
}
\value{
Downloads a compressed \code{sqlite} file to the current working
directory. The file is named \code{cRegulome.db.gz} by default and it's
not advised to change the name to avoid breaking the other functions
that calls the database.
}
\description{
This function calls \code{\link[utils]{download.file}} to download the
pre-build database file of cRegulome. Additionally, the function checks
the validity of the pre-defined URL and whether the database file exists
in the current working directory to avoid redownloading it. Typically,
users would run this function once at the first time the use the package
or to update the database to the latest version.
}
\examples{
\dontrun{
# downlaod db file
get_db()

# check it exits in the current working directory
# should return TRUE
file.exists('./cRegulome.db.gz')

# download a smaller test db file
get_db(test = TRUE)
}

# load the test db file from shipped with the pacakge
db_file <- devtools::system.file("extdata", "cRegulome.db", package = "cRegulome")
file.info(db_file)

}
