% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_mir}
\alias{get_mir}
\title{Get microRNA correlations from cRegulome.db}
\usage{
get_mir(conn, mir, study = NULL, min_cor = NULL, max_num = NULL,
  targets_only = FALSE)
}
\arguments{
\item{conn}{A connection to the database file by \code{\link[DBI]{dbConnect}}}

\item{mir}{A required \code{character} vector of the microRNAs of interest.
These are the miRBase ID which are the official identifiers of the
widely used miRBase database, \url{http://www.mirbase.org/}.}

\item{study}{A \code{character} vector of The Cancer Genome Atlase (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
defult \code{NULL} all available studies will be included.}

\item{min_cor}{A \code{numeric}, an absoute correlation minimmum between 0
and 1 for each \code{mir}.}

\item{max_num}{An \code{integer}, maximum number of \code{features} to show
for each \code{mir} in each \code{study}.}

\item{targets_only}{A \code{logical}, default \code{FALSE}. When
\code{TRUE}, \code{features} will be the microRNA targets as defined in
the package \code{\link[targetscan.Hs.eg.db]{targetscan.Hs.eg.db}}.}
}
\value{
A tidy \code{data.frame} of four columns. \code{mirna_base} is the
microRNA miRBase IDs, \code{feature} is the features/genes, \code{cor}
is the corresponding expression correaltions and \code{study} is TCGA
study ID.
}
\description{
This function access the \code{sqlite} database file which is obtained by
running \link{get_db}.The function returns an error if the uncompressd
database file \code{cRegulome.db} is not in the working directory.
Basically, the function provides fileters to subset the large tables to
the items of interest.
}
\examples{
\dontrun{
# downlaod db file
get_db(test = TRUE)
gunzip('cRegulome.db.gz')
}
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# provide only required arguments
dat <- get_mir(conn, mir = 'hsa-let-7b')

# optional arguments
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'),
    min_cor = .5,
    max_num = 100,
    targets_only = TRUE)
DBI::dbDisconnect(conn)

}
