% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_tf}
\alias{get_tf}
\title{Get transcription factor correlations from cRegulome.db}
\usage{
get_tf(tf, study = NULL, min_cor = NULL, max_num = NULL,
  targets_only = FALSE)
}
\arguments{
\item{tf}{A required \code{character} vector of the transcription factor of
interest. These are the official gene symbols of the genes contains the
transcription factor.}

\item{study}{A \code{character} vector of The Cancer Genome Atlase (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
 defult \code{NULL} all available studies will be included.}

\item{min_cor}{A \code{numeric}, an absoute correlation minimmum between 0
and 1 for each \code{tf}.}

\item{max_num}{An \code{integer}, maximum number of \code{features} to show
for each \code{tf} in each \code{study}.}

\item{targets_only}{A \code{logical}, default \code{FALSE}. When
\code{TRUE}, \code{features} will be the targets of the transcription
factors as identified in the Cistrome Cancer,
\url{http://cistrome.org/CistromeCancer/}}
}
\value{
A tidy \code{data.frame} of four columns. \code{tf} is the official
   gene symbols of the genes contains the transcription factor,
   \code{feature} is the features/genes, \code{cor} is the corresponding
   expression correaltions and \code{study} is TCGA study ID.
}
\description{
This function access the \code{sqlite} database file which is obtained by
running \link{get_db}. The function returns an error if the uncompressd
database file \code{cRegulome.db} is not in the working directory.
Basically, the function provides fileters to subset the large tables to the
items of interest.
}
\examples{
# downlaod db file
get_db(test = TRUE)

# provide only required argument
dat <- get_tf('AFF4')

# optional arguments
dat <- get_mir(c('AFF4', 'ESR1'),
    study = c('ACC', 'BLCA'),
    min_cor = .5,
    max_num = 100,
    targets_only = TRUE)
}
