% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{mir}
\alias{mir}
\title{Published list of microNAs related to gastric cancer}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 93 rows and 15 columns.}
\usage{
mir
}
\description{
Published list of microNAs related to gastric cancer
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4393113/bin/pone.0122882.s003.xls}
}
\keyword{datasets}
