% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{mirCancer}
\alias{mirCancer}
\title{miRCancerdb query for 49 microRNAs}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 299324 rows and 4 columns.}
\usage{
mirCancer
}
\description{
miRCancerdb query for 49 microRNAs
}
\references{
\url{https://www.dropbox.com/s/89hr8t8mcpcec39/mircancer.csv?raw=1}
}
\keyword{datasets}
