% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\alias{print.cmicroRNA}
\alias{print.cTF}
\title{Print method for \link{cmicroRNA} and \link{cTF} objects}
\usage{
print(ob, ...)

\method{print}{cmicroRNA}(ob, ...)

\method{print}{cTF}(ob, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by calling \link{cmicroRNA} or \link{cTF}.}

\item{...}{Other argument to \code{\link[base]{print}}.}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Print method for \link{cmicroRNA} and \link{cTF} objects
}
\examples{
\dontrun{
# get data for 2 microRNAs in the ACC study
dat <- get_mir(c('hsa-let-7b', 'hsa-mir-134'),
 study = c('ACC', 'BLCA'),
 min_cor = .5)

# convert to cmicroRNA object and plot
ob <- cmicroRNA(dat)
print(ob)

#' # get data for 2 transcription factors in the ACC study
dat <- get_tf(c('AFF4', 'ESR1'),
 study = c('ACC', 'BLCA'),
 min_cor = .5)

# convert to cTF object and plot
ob <- cTF(dat)
print(ob)
}
}
