% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{stat_collect}
\alias{stat_collect}
\title{Collect data from SQLite database}
\usage{
stat_collect(conn, study, stat, type = "mir")
}
\arguments{
\item{conn}{A connection such as this returned by 
\code{\link[DBI]{dbConnect}}}

\item{study}{A \code{character} vector of The Cancer Genome Atlas (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
default \code{NULL} all available studies will be included.}

\item{stat}{A string such as this returned by \code{\link{stat_make}}}

\item{type}{A \code{character} string. Either 'mir' of 'tf'. Used to define
columns and tables names.}
}
\value{
A data.frame
}
\description{
Not meant to be called direclty by the user.
}
