% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{stat_collect_targets}
\alias{stat_collect_targets}
\title{Collect target features from SQLite database}
\usage{
stat_collect_targets(conn, stat)
}
\arguments{
\item{conn}{A connection such as this returned by 
\code{\link[DBI]{dbConnect}}}

\item{stat}{A string such as this returned by \code{\link{stat_make}}}
}
\value{
A \code{character} vector
}
\description{
Not meant to be called directly by the user.
}
