% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{stat_make_targets}
\alias{stat_make_targets}
\title{Make A SQL statement to extract target features}
\usage{
stat_make_targets(reg, study, type = "mir")
}
\arguments{
\item{reg}{A \code{character} vector of one or more regulator ID.}

\item{study}{A \code{character} vector of The Cancer Genome Atlas (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
default \code{NULL} all available studies will be included.}

\item{type}{A \code{character} string. Either 'mir' of 'tf'. Used to define
columns and tables names.}
}
\value{
A character string
}
\description{
Not meant to be called directly by the user.
}
