% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_objects.R
\docType{data}
\name{tf}
\alias{tf}
\title{Published list of transcription factors related to gastric cancer}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 2 columns.}
\usage{
tf
}
\description{
Published list of transcription factors related to gastric cancer
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4393113/bin/pone.0122882.s001.xlsx}
}
\keyword{datasets}
