% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{venn.diagram}
\alias{venn.diagram}
\title{Venn Diagram of microRNA or transcription factor correlated features}
\usage{
venn.diagram(ob, study = NULL, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{study}{A \code{character} vector of The Cancer Genome Atlase (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
defult \code{NULL} all available studies will be included.}

\item{...}{Other options}
}
\value{
A venn diagram with a ciccle or an ellipses for each microRNA and
the number of correlated features.
}
\description{
Count and plot the numbers of microRNA correlated features in
\code{cmicroRNA} object.
}
\examples{
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# get data for 2 microRNAs in the ACC study
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'),
    min_cor = .5)
DBI::dbDisconnect(conn)

# convert to cmicroRNA object and plot
ob <- cmicroRNA(dat)
venn.diagram(ob, study = 'ACC')

}
