% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cmicroRNA}
\alias{cmicroRNA}
\title{Construct cmicroRNA object}
\usage{
cmicroRNA(dat_mir)
}
\arguments{
\item{dat_mir}{A \code{data.frame} such as this returned by calling \link{get_mir}.}
}
\value{
An S3 object of class \code{cmicroRNA}
}
\description{
Constructs an S3 object called cmicroRNA contains data returned by calling \link{get_mir}. Used to define
methods for printing and visualizing microRNA-gene expression correlations.
}
\examples{
\dontrun{
# get data for 2 microRNAs in the ACC study
dat <- get_mir(c('hsa-let-7b', 'hsa-mir-134'),
 study = c('ACC', 'BLCA'),
 min_cor = .5)

# convert to cmicroRNA object
ob <- cmicroRNA(dat)
}
}
