% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cmicroRNA}
\alias{cmicroRNA}
\title{Construct cmicroRNA object}
\usage{
cmicroRNA(dat_mir)
}
\arguments{
\item{dat_mir}{A \code{data.frame} such as this returned by calling
\link{get_mir}.}
}
\value{
An S3 object of class \code{cmicroRNA}
}
\description{
Constructs an S3 object called cmicroRNA contains data returned by calling
\link{get_mir}. Used to define methods for printing and visualizing
microRNA-gene expression correlations.
}
\examples{
# connect to test database file
db_file <- devtools::system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# optional arguments
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'),
    min_cor = .5,
    max_num = 100,
    targets_only = TRUE)
DBI::dbDisconnect(conn)

# convert object
ob <- cmicroRNA(dat)

}
