% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cor_print}
\alias{cor_print}
\title{Print method for \link{cmicroRNA} and \link{cTF} objects}
\usage{
cor_print(ob, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{...}{Other argument to \code{\link[base]{print}}.}
}
\value{
Printed text
}
\description{
Print method for \link{cmicroRNA} and \link{cTF} objects
}
\examples{
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# get data for 2 microRNAs in the ACC study
dat <- get_mir(conn,
    mir = c('hsa-let-7b', 'hsa-mir-134'),
    study = c('ACC', 'BLCA'),
    min_abs_cor = .5)
DBI::dbDisconnect(conn)

# convert to cmicroRNA object
ob <- cmicroRNA(dat)
cor_print(ob)

}
