% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.plots.R
\name{cor_upset}
\alias{cor_upset}
\title{\code{\link[UpSetR]{upset}} plot of microRNA or tf sets}
\usage{
cor_upset(ob, study, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{study}{A \code{character} vector of The Cancer Genome Atlas (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
default \code{NULL} all available studies will be included.}

\item{...}{Other options}
}
\value{
An \code{\link[UpSetR]{upset}} plot
}
\description{
\code{\link[UpSetR]{upset}} of sets of microRNAs or transcription
factors and their correlated features in a TCGA study.
}
\examples{
 
# locate the testset file and connect
fl <- system.file('extdata', 'cRegulome.db', package = 'cRegulome')
conn <- RSQLite::dbConnect(RSQLite::SQLite(), fl)

# enter a custom query with different arguments
dat <- get_mir(conn,
               mir = c('hsa-let-7g', 'hsa-let-7i'),
               study = 'STES')

# make a cmicroRNA object   
cmir <- cmicroRNA(dat)

# print object
cor_upset(cmir)

}
