% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cRegulome.R
\docType{package}
\name{cRegulome}
\alias{cRegulome}
\alias{cRegulome-package}
\title{\code{cRegulome} package}
\description{
Download, access and visualize Regulome (microRNA and transcription factors)
data from miRCancer and Cistrome cancer
}
