% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cTF}
\alias{cTF}
\title{Construct cTF object}
\usage{
cTF(dat_tf)
}
\arguments{
\item{dat_tf}{A \code{data.frame} such as this returned by calling \link{get_tf}.}
}
\value{
An S3 object of class \code{cTF}
}
\description{
Constructs an S3 object called cTF contains data returned by calling \link{get_tf}. Used to define
methods for printing and visualizing transcription factors-gene expression correlations.
}
\examples{
# downlaod db file
get_db(test = TRUE)

# get data for 2 tarnscription factors in the ACC study
dat <- get_tf(c('AFF4', 'ESR1'),
 study = c('ACC', 'BLCA'),
 min_cor = .5)

# convert to cTF object
ob <- cTF(dat)

}
