% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cTF}
\alias{cTF}
\title{Construct cTF object}
\usage{
cTF(dat_tf)
}
\arguments{
\item{dat_tf}{A \code{data.frame} such as this returned by calling
\link{get_tf}.}
}
\value{
An S3 object of class \code{cTF}
}
\description{
Constructs an S3 object called cTF contains data returned by calling
\link{get_tf}. Used to define methods for printing and visualizing
transcription factors-gene expression correlations.
}
\examples{
# connect to test database file
db_file <- system.file("extdata", "cRegulome.db", package = "cRegulome")
conn <- DBI::dbConnect(RSQLite::SQLite(), db_file)

# optional arguments
dat <- get_tf(conn,
    tf = c('AFF4', 'ESR1'),
    study = c('ACC', 'BLCA'),
    min_abs_cor = .5,
    max_num = 100,
    targets_only = TRUE)
DBI::dbDisconnect(conn)

# convert to object
ob <- cTF(dat)

}
