% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot}
\alias{plot}
\title{Plot method for \link{cmicroRNA} and \link{cTF} objects}
\usage{
plot(ob, study = NULL, ...)
}
\arguments{
\item{ob}{A \link{cmicroRNA} or \link{cTF} object such as this returned by
calling \link{cmicroRNA} or \link{cTF}.}

\item{study}{A \code{character} vector of The Cancer Genome Atlase (TCGA)
study identifiers. To view the available studies in TCGA project,
\url{https://tcga-data.nci.nih.gov/docs/publications/tcga}. When left to
defult \code{NULL} all available studies will be included.}

\item{...}{Other options}
}
\value{
A \code{ggplot} object of dot plot.
}
\description{
A dot plot of microRNA/TF correlation in a single study of TCGA. When the
object \link{cmicroRNA}/\link{cTF} contains more than one TCGA studies, the
argument \code{study} is a requirment.
}
